Refactor more stuff into helpers.rs
