+++ crates/icondiffbot2/Cargo.toml
@@ -23,6 +23,10 @@ flume = "0.10.12"
 ndarray = "0.15.3"
 sanitize-filename = "0.4.0"
 
+[dependencies.tracing]
+version = "0.1.35"
+features = ["attributes"]
+
 [dependencies.rocket]
 version = "0.5.0-rc.1"
 features = ["json"]
